nixos/networking: Add the FQDN and hostname to /etc/hosts
This fixes the output of "hostname --fqdn" (previously the domain name
was not appended). Additionally it's now possible to use the FQDN.

This works by unconditionally adding two entries to /etc/hosts:
127.0.0.1 localhost
::1 localhost

These are the first two entries and therefore gethostbyaddr() will
always resolve "127.0.0.1" and "::1" to "localhost".

The FQDN and hostname entries are appended later to /etc/hosts, e.g.:
127.0.0.1 nixos-unstable.test.tld nixos-unstable
::1 nixos-unstable.test.tld nixos-unstable

Therefore gethostbyname() resolves "nixos-unstable" to the FQDN
(canonical name): "nixos-unstable.test.tld".

Advantages to the previous behaviour:
- The hostname will now resolve to "127.0.0.1" instead of "127.0.1.1" while
  we still avoid the issue mentioned in c57892462bf. The old behaviour
  can cause issues as we have services listening ONLY to "127.0.0.1" or
  "localhost" by default which means that connecting to the hostname
  (i.e. "127.0.1.1") will not work!
- The FQDN will now also be resolved correctly (the entry was missing).
- E.g. the command "hostname --fqdn" will now work as expected.
Drawbacks:
- Unknown (could potentially cause other unexpected issues).